backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn 

sqlalchemy

bcrypt

python-dotenv

pymongo

cloudinary

google-auth

python-multipart

sympy

requests

python-jose[cryptography]


